requirements.txt
altgraph==0.17.4
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
defusedxml==0.7.1
et_xmlfile==2.0.0
fonttools==4.55.5
fpdf2==2.8.2
greenlet==3.1.1
idna==3.10
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
pycparser==2.22
pydub==0.25.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sounddevice==0.5.1
SQLAlchemy==2.0.37
srt==3.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vosk==0.3.45
websockets==14.2
